fix: project bundle file artifacts in ability lifecycle
