+++ inc/Engine/Bundle/AgentBundleAbilityService.php
@@ -24,12 +24,14 @@ final class AgentBundleAbilityService {
 	private Pipelines $pipelines;
 	private Flows $flows;
 	private AgentBundler $bundler;
+	private AgentBundleLifecycleProjection $projection;
 
-	public function __construct( ?Agents $agents = null, ?Pipelines $pipelines = null, ?Flows $flows = null, ?AgentBundler $bundler = null ) {
+	public function __construct( ?Agents $agents = null, ?Pipelines $pipelines = null, ?Flows $flows = null, ?AgentBundler $bundler = null, ?AgentBundleLifecycleProjection $projection = null ) {
 		$this->agents    = $agents ?? new Agents();
 		$this->pipelines = $pipelines ?? new Pipelines();
 		$this->flows     = $flows ?? new Flows();
 		$this->bundler   = $bundler ?? new AgentBundler();
+		$this->projection = $projection ?? new AgentBundleLifecycleProjection( $this->pipelines, $this->flows );
 	}
 
 	/** @return array<string,mixed> */
@@ -191,7 +193,7 @@ final class AgentBundleAbilityService {
 				array(
 					'bundle'             => $this->bundle_summary( $bundle, $slug ),
 					'agent'              => $agent ? $agent : array(),
-					'target_artifacts'   => $this->bundle_artifacts( $bundle ),
+					'target_artifacts'   => $this->projection->target_artifacts( $bundle, $agent ?: null ),
 					'approved_artifacts' => $this->approved_rebased_artifact_keys( $rebased_artifacts ),
 					'rebased_artifacts'  => $rebased_artifacts,
 					'summary'            => 'Review locally modified bundle artifacts before applying.',
@@ -290,7 +292,7 @@ final class AgentBundleAbilityService {
 	private function plan_for_bundle( array $bundle, string $slug = '' ): AgentBundleUpgradePlan {
 		$agent = $this->resolve_bundle_agent( $bundle, $slug );
 		if ( ! $agent ) {
-			return AgentBundleUpgradePlanner::plan( array(), array(), $this->bundle_artifacts( $bundle ), $this->bundle_summary( $bundle, $slug ) );
+			return AgentBundleUpgradePlanner::plan( array(), array(), $this->projection->target_artifacts( $bundle ), $this->bundle_summary( $bundle, $slug ) );
 		}
 
 		$bundle_state = is_array( $agent['agent_config']['datamachine_bundle'] ?? null ) ? $agent['agent_config']['datamachine_bundle'] : array();
@@ -298,138 +300,10 @@ final class AgentBundleAbilityService {
 
 		return AgentBundleUpgradePlanner::plan(
 			$installed,
-			$this->current_artifacts( $agent, $installed ),
-			$this->bundle_artifacts_for_agent( $bundle, $agent ),
+			$this->projection->current_artifacts( $agent, $installed ),
+			$this->projection->target_artifacts( $bundle, $agent ),
 			$this->bundle_summary( $bundle, $slug )
 		);
-	}
-
-	/** @return array<int,array<string,mixed>> */
-	private function bundle_artifacts( array $bundle ): array {
-		return $this->bundle_artifacts_for_agent( $bundle, null );
-	}
-
-	/** @return array<int,array<string,mixed>> */
-	private function bundle_artifacts_for_agent( array $bundle, ?array $agent ): array {
-		$artifacts                  = array();
-		$agent_id                   = is_array( $agent ) ? (int) ( $agent['agent_id'] ?? 0 ) : 0;
-		$pipeline_id_map            = array();
-		$existing_pipelines_by_slug = array();
-		$incoming_config            = is_array( $bundle['agent']['agent_config'] ?? null ) ? $bundle['agent']['agent_config'] : array();
-
-		$artifacts[] = array(
-			'artifact_type' => 'agent_config',
-			'artifact_id'   => 'config',
-			'source_path'   => 'manifest.json#/agent/agent_config',
-			'payload'       => AgentBundleAgentConfig::tracked_payload( $incoming_config ),
-		);
-
-		if ( $agent_id > 0 ) {
-			foreach ( $this->pipelines->get_all_pipelines( null, $agent_id ) as $pipeline ) {
-				$existing_pipelines_by_slug[ (string) ( $pipeline['portable_slug'] ?? '' ) ] = $pipeline;
-			}
-		}
-
-		foreach ( $bundle['pipelines'] ?? array() as $pipeline ) {
-			if ( ! is_array( $pipeline ) ) {
-				continue;
-			}
-			$slug = PortableSlug::normalize( (string) ( $pipeline['portable_slug'] ?? ( $pipeline['pipeline_name'] ?? 'pipeline' ) ), 'pipeline' );
-			if ( isset( $existing_pipelines_by_slug[ $slug ] ) ) {
-				$old_id                      = (int) ( $pipeline['original_id'] ?? 0 );
-				$new_id                      = (int) ( $existing_pipelines_by_slug[ $slug ]['pipeline_id'] ?? 0 );
-				$pipeline_id_map[ $old_id ]  = $new_id;
-				$pipeline['pipeline_config'] = BundleStepIdRemapper::remap_pipeline_step_ids( is_array( $pipeline['pipeline_config'] ?? null ) ? $pipeline['pipeline_config'] : array(), $old_id, $new_id );
-			}
-
-			$artifacts[] = array(
-				'artifact_type' => 'pipeline',
-				'artifact_id'   => $slug,
-				'source_path'   => 'pipelines/' . $slug . '.json',
-				'payload'       => $this->pipeline_payload( $pipeline, $slug ),
-			);
-		}
-
-		foreach ( $bundle['flows'] ?? array() as $flow ) {
-			if ( ! is_array( $flow ) ) {
-				continue;
-			}
-			$slug            = PortableSlug::normalize( (string) ( $flow['portable_slug'] ?? ( $flow['flow_name'] ?? 'flow' ) ), 'flow' );
-			$old_pipeline_id = (int) ( $flow['original_pipeline_id'] ?? 0 );
-			$new_pipeline_id = (int) ( $pipeline_id_map[ $old_pipeline_id ] ?? 0 );
-			$existing_flow   = $new_pipeline_id > 0 ? $this->flows->get_by_portable_slug( $new_pipeline_id, $slug ) : null;
-
-			if ( $existing_flow ) {
-				$flow['flow_config'] = BundleStepIdRemapper::remap_flow_step_ids( is_array( $flow['flow_config'] ?? null ) ? $flow['flow_config'] : array(), $old_pipeline_id, $new_pipeline_id, (int) $existing_flow['flow_id'] );
-			}
-
-			$artifacts[] = array(
-				'artifact_type' => 'flow',
-				'artifact_id'   => $slug,
-				'source_path'   => 'flows/' . $slug . '.json',
-				'payload'       => $this->flow_payload( $flow, $slug ),
-			);
-		}
-
-		foreach ( AgentBundleArtifactExtensions::normalize_artifacts( is_array( $bundle['extension_artifacts'] ?? null ) ? $bundle['extension_artifacts'] : array() ) as $artifact ) {
-			$artifacts[] = $artifact;
-		}
-
-		return $artifacts;
-	}
-
-	/** @param array<int,array<string,mixed>> $installed */
-	private function current_artifacts( array $agent, array $installed ): array {
-		$agent_id  = (int) $agent['agent_id'];
-		$artifacts = array();
-		$pipelines = $this->pipelines->get_all_pipelines( null, $agent_id );
-		$flows     = $this->flows->get_all_flows( null, $agent_id );
-
-		$artifacts[] = array(
-			'artifact_type' => 'agent_config',
-			'artifact_id'   => 'config',
-			'source_path'   => 'manifest.json#/agent/agent_config',
-			'payload'       => AgentBundleAgentConfig::tracked_payload( is_array( $agent['agent_config'] ?? null ) ? $agent['agent_config'] : array() ),
-		);
-
-		$pipeline_by_slug = array();
-		foreach ( $pipelines as $pipeline ) {
-			$slug = (string) ( $pipeline['portable_slug'] ?? '' );
-			if ( '' !== $slug ) {
-				$pipeline_by_slug[ $slug ] = $pipeline;
-			}
-		}
-
-		$flow_by_slug = array();
-		foreach ( $flows as $flow ) {
-			$slug = (string) ( $flow['portable_slug'] ?? '' );
-			if ( '' !== $slug ) {
-				$flow_by_slug[ $slug ] = $flow;
-			}
-		}
-
-		foreach ( $installed as $record ) {
-			$type = (string) ( $record['artifact_type'] ?? '' );
-			$id   = (string) ( $record['artifact_id'] ?? '' );
-			if ( 'pipeline' === $type && isset( $pipeline_by_slug[ $id ] ) ) {
-				$artifacts[] = array(
-					'artifact_type' => 'pipeline',
-					'artifact_id'   => $id,
-					'source_path'   => (string) ( $record['source_path'] ?? '' ),
-					'payload'       => $this->pipeline_payload( $pipeline_by_slug[ $id ], $id ),
-				);
-			}
-			if ( 'flow' === $type && isset( $flow_by_slug[ $id ] ) ) {
-				$artifacts[] = array(
-					'artifact_type' => 'flow',
-					'artifact_id'   => $id,
-					'source_path'   => (string) ( $record['source_path'] ?? '' ),
-					'payload'       => $this->flow_payload( $flow_by_slug[ $id ], $id ),
-				);
-			}
-		}
-
-		return array_merge( $artifacts, AgentBundleArtifactExtensions::current_artifacts( $agent, $installed, array( 'agent_id' => $agent_id ) ) );
 	}
 
 	private function add_runtime_drift_to_plan( array $plan, array $bundle, string $slug = '', string $decision = 'preserve_existing' ): array {
@@ -525,13 +399,13 @@ final class AgentBundleAbilityService {
 		}
 
 		$current_index = array();
-		foreach ( $this->current_artifacts( $agent, $installed ) as $artifact ) {
+		foreach ( $this->projection->current_artifacts( $agent, $installed ) as $artifact ) {
 			$key                   = AgentBundleArtifactExtensions::artifact_key( (string) $artifact['artifact_type'], (string) $artifact['artifact_id'] );
 			$current_index[ $key ] = $artifact;
 		}
 
 		$target_index = array();
-		foreach ( $this->bundle_artifacts_for_agent( $bundle, $agent ) as $artifact ) {
+		foreach ( $this->projection->target_artifacts( $bundle, $agent ) as $artifact ) {
 			$key                  = AgentBundleArtifactExtensions::artifact_key( (string) $artifact['artifact_type'], (string) $artifact['artifact_id'] );
 			$target_index[ $key ] = $artifact;
 		}
@@ -587,42 +461,6 @@ final class AgentBundleAbilityService {
 		}
 
 		return array_values( array_unique( $approved ) );
-	}
-
-	private function pipeline_payload( array $pipeline, string $portable_slug ): array {
-		return array(
-			'portable_slug'   => $portable_slug,
-			'pipeline_name'   => (string) ( $pipeline['pipeline_name'] ?? '' ),
-			'pipeline_config' => is_array( $pipeline['pipeline_config'] ?? null ) ? $pipeline['pipeline_config'] : array(),
-		);
-	}
-
-	private function flow_payload( array $flow, string $portable_slug ): array {
-		return array(
-			'portable_slug'     => $portable_slug,
-			'flow_name'         => (string) ( $flow['flow_name'] ?? '' ),
-			'flow_config'       => $this->flow_config_without_runtime_queues( is_array( $flow['flow_config'] ?? null ) ? $flow['flow_config'] : array() ),
-			'scheduling_policy' => $this->flow_scheduling_policy( is_array( $flow['scheduling_config'] ?? null ) ? $flow['scheduling_config'] : array() ),
-			'queue_policy'      => 'create_seed_upgrade_preserve_existing',
-		);
-	}
-
-	private function flow_scheduling_policy( array $config ): string {
-		$interval = (string) ( $config['interval'] ?? 'manual' );
-		$enabled  = array_key_exists( 'enabled', $config ) ? false !== $config['enabled'] : 'manual' !== $interval;
-
-		return ( 'manual' === $interval || ! $enabled ) ? 'create_paused_upgrade_preserve_existing' : 'create_bundle_schedule_upgrade_preserve_existing';
-	}
-
-	private function flow_config_without_runtime_queues( array $flow_config ): array {
-		foreach ( $flow_config as &$step ) {
-			if ( is_array( $step ) ) {
-				unset( $step['prompt_queue'], $step['config_patch_queue'], $step['queue_mode'], $step['_queue_consume_revision'] );
-			}
-		}
-		unset( $step );
-
-		return $flow_config;
 	}
 
 	private function resolve_bundle_agent( array $bundle, string $slug = '' ): ?array {
